fix: change price_at_time type from float to Decimal in OrderItem

+++ backend/schemas/cart_schema.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field
 from typing import List
 from uuid import UUID, uuid4
@@ -5,7 +7,7 @@ from uuid import UUID, uuid4
 class OrderItem(BaseModel):
     menu_item_id: UUID = Field(default_factory=uuid4)
     quantity: int
-    price_at_time: float
+    price_at_time: Decimal
 
 class Cart(BaseModel):
     customer_id: str
